% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adhdData2}
\alias{adhdData2}
\title{Children's reaction times (milliseconds) to stimuli of different nature, arranged with one single
response column and taking the multi-variate response as an explicit within-subjects factor.}
\format{A data frame with 120 rows and 4 variables:
\describe{
   \item{Group}{whether the child has ADHD or is healty (normal)}.
   \item{Stimulus}{the stimulus to which the reaction time in this row corresponds.}
   \item{Subject}{an integer ID that identifies the subject to which the reaction time corresponds.}
   \item{Milliseconds}{reaction time (milliseconds) to of the aforementioned subject to the aforementioned stimulus.}
}}
\usage{
adhdData2
}
\description{
Exactly the same data explained in "adhdData" but reshaped as follows.
}
\keyword{datasets}
