% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welchADFt-methods.R
\name{summary.welchADFt}
\alias{summary.welchADFt}
\alias{print.welchADFt}
\title{Summarizing Welch Approximate Degrees of Freedom test results}
\usage{
\method{summary}{welchADFt}(object, verbose = FALSE, digits = max(4,
  getOption("digits")), ...)

\method{print}{welchADFt}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{object}{an object of class "welchADFt" returned by a call to 
\code{\link{welchADF.test}}.}

\item{verbose}{whether the summary table should be preceded by an extended description of the
welch ADF test performed, including the kind of means (trimmed or not), kind of test
(omnibus or pairwise) and whether bootstrap was used or not. Defaults to FALSE.}

\item{digits}{the number of significant digits to use when printing.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{an object of class "welchADFt".}
}
\value{
a string which summarizes the info, with class \code{"summary.welchADFt"}
  ready to be printed by specific method \code{print.summary.welchADFt}.
}
\description{
\code{summary} and \code{print} methods for class \code{"welchADF"}
}
