% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miceData}
\alias{miceData}
\title{Number of visits and time spent in different tunnels of laboratory mice}
\format{A data frame with 144 rows and 6 variables:
\describe{
   \item{Subject}{an ID that identifies the mouse to which the measurements of the row correspond.}
   \item{nurs}{type of nursing mother this mouse had at birth.}
   \item{tunnel}{type of tunnel to which the measurements of this row correspond.}
   \item{visits}{number of visits to the aforementioned tunnel.}
   \item{time}{total time spent by the mouse at this tunnel.}
   \item{latency}{seconds passed before the mouse first visited this tunnel.}
}}
\source{
\url{http://core.ecu.edu/psyc/wuenschk/SPSS/TUNNEL4b.sav}
}
\usage{
miceData
}
\description{
Wild strain house mice were, at birth, cross fostered onto house mouse (Mus), 
deer mouse (Peromyscus) or rat (Rattus) nursing mothers. Ten days after weaning, 
each subject was tested in an apparatus that allowed it to enter four different tunnels: 
one scented with clean pine shavings, and the other three tunnels with shavings 
bearing the scent of Mus, Peromyscus, or Rattus respectively. 
Three variables were measured for each tunnel: the number of visits to the 
tunnel during a twenty minute test, the time spent by each subject
in each of the four tunnels and the latency to first visit of each tunnel.
}
\references{
K. L.Wuensch. Fostering house mice onto rats and deer mice: Effects on response to species odors.
Animal Learning & Behavior, 20(3):253 - 258, 1992
}
\keyword{datasets}
