% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{perceptionData}
\alias{perceptionData}
\title{Number of puzzles 42 students were able to solve}
\format{A data frame with 42 rows and 2 variables:
\describe{
  \item{Group}{group of the student (what the student was asked to imagine)}
  \item{y}{number of puzzles the student was able to solve, out of 12}
}}
\usage{
perceptionData
}
\description{
An artificial dataset created by Lix et al. which recreates data reported by a real study
on perception and concentration, on which 42 students were given several puzzles to be solved. 
The students are divided into three balanced groups as they had previously been asked to 
imagine solving puzzles in the distant future, near future, or not to imagine anything at all 
(control group).
}
\examples{
omnibus_LSM <- welchADF.test(perceptionData, response = "y", between.s = "Group")
omnibus_trimmed <- update(omnibus_LSM, trimming = TRUE)
omnibus_trimmed_boot <- update(omnibus_trimmed, bootstrap = TRUE, seed = 12345, numsim_b = 600)
pairwise_LSM <- update(omnibus_LSM, effect = "Group", contrast = "all.pairwise")
pairwise_trimmed <- update(pairwise_LSM, trimming = TRUE, effect.size = TRUE)
pairwise_trimmed_boot <- update(pairwise_trimmed, bootstrap = TRUE, seed = 12345, numsim_b = 600)
summary(omnibus_LSM)
summary(pairwise_trimmed_boot, digits = 6) # digits defaults to max(4, getOption("digits"))
}
\references{
Forster, J., Liberman, N., & Friedman, R.S. (2004). Temporal construal effects on
abstract and concrete thinking: consequences for insight and creative cognition.
Journal of Personality and Social Psychology, 87, 2, 177-189.
}
\keyword{datasets}
